Fixes so compass can compile locally
Remove references to a number of undefined variables and mixins, so that
the compass task in the local Gruntfile can complete its job
successfully.

I don't expect that a fully-compiled Quilt css and js file will be
essential to any projects in the way it was with Quilt 2.x, but it
should still be available, just because. Also, being able to run a grunt
compass task over this repo provides a better development workflow and
should catch some obvious bugs before this gets introduced into other
work.